requirements_modular.txt
streamlit>=1.28.0
psycopg2-binary==2.9.9
pandas>=1.3.0
pillow>=8.0.0
plotly>=5.0.0
numpy>=1.21.0
streamlit-webrtc>=0.45.0
opencv-python-headless>=4.5.5
pyzbar>=0.1.8
av>=10.0.0